refactor: move feature sections out of shared components
