+++ src/features/home/components/Carousel.tsx
