+++ src/features/work/components/WorkIndexPage.tsx
@@ -1,12 +1,12 @@
-import SelectedWork from "@/components/SelectedWork";
 import { getWorkSummaries } from "@/lib/work";
+import SelectedWorkSection from "@/features/work/components/SelectedWorkSection";
 
 export default async function WorkIndexPage() {
   const workItems = await getWorkSummaries();
 
   return (
     <main className="bg-neutral-950">
-      <SelectedWork
+      <SelectedWorkSection
         workItems={workItems}
         heading="All Work & Case Studies"
         description="A deeper look at the products and case studies I’ve delivered—from mobile apps to full-stack builds."
